Restructure batterysync/mobile
