+++ batterysync/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/data/repository/BatteryStatsDbRepository.kt
@@ -1,6 +1,8 @@
-package com.boswelja.smartwatchextensions.batterysync
+package com.boswelja.smartwatchextensions.batterysync.data.repository
 
+import com.boswelja.smartwatchextensions.batterysync.BatteryStats
 import com.boswelja.smartwatchextensions.batterysync.database.BatteryStatsDatabase
+import com.boswelja.smartwatchextensions.batterysync.domain.repository.BatteryStatsRepository
 import com.squareup.sqldelight.runtime.coroutines.asFlow
 import com.squareup.sqldelight.runtime.coroutines.mapToOneOrNull
 import kotlinx.coroutines.flow.Flow
