+++ batterysync/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/di/FeatureDI.kt
@@ -1,5 +1,7 @@
 package com.boswelja.smartwatchextensions.batterysync
 
+import com.boswelja.smartwatchextensions.batterysync.data.repository.BatteryStatsDbRepository
+import com.boswelja.smartwatchextensions.batterysync.domain.repository.BatteryStatsRepository
 import com.boswelja.smartwatchextensions.batterysync.database.BatteryStatsDatabase
 import org.koin.core.parameter.parametersOf
 import org.koin.core.qualifier.named
