+++ batterysync/mobile/src/main/sqldelight/com/boswelja/smartwatchextensions/batterysync/data/database/WatchBatteryStats.sq
