+++ batterysync/mobile/src/test/kotlin/com/boswelja/smartwatchextensions/batterysync/BatteryStatsDbRepositoryTest.kt
@@ -1,6 +1,8 @@
 package com.boswelja.smartwatchextensions.batterysync
 
 import app.cash.turbine.test
+import com.boswelja.smartwatchextensions.batterysync.data.repository.BatteryStatsDbRepository
+import com.boswelja.smartwatchextensions.batterysync.domain.repository.BatteryStatsRepository
 import com.boswelja.smartwatchextensions.batterysync.database.BatteryStatsDatabase
 import com.squareup.sqldelight.db.SqlDriver
 import com.squareup.sqldelight.sqlite.driver.JdbcSqliteDriver
