+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/dashboard/ui/DashboardViewModel.kt
@@ -3,7 +3,7 @@ package com.boswelja.smartwatchextensions.dashboard.ui
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.boswelja.smartwatchextensions.appmanager.WatchAppRepository
-import com.boswelja.smartwatchextensions.batterysync.BatteryStatsRepository
+import com.boswelja.smartwatchextensions.batterysync.domain.repository.BatteryStatsRepository
 import com.boswelja.smartwatchextensions.core.devicemanagement.SelectedWatchManager
 import com.boswelja.smartwatchextensions.core.devicemanagement.WatchRepository
 import com.boswelja.watchconnection.common.Watch
